RcppParallel-package {RcppParallel}	R Documentation

Parallel programming tools for Rcpp

Description

High level functions for doing parallel programming with Rcpp. For example, the parallelFor function can be used to convert the work of a standard serial "for" loop into a parallel one and the parallelReduce function can be used for accumulating aggregate or other values.

The high level interface enables safe and robust parallel programming without direct manipulation of operating system threads. On Windows, OS X, and Linux systems the underlying implementation is based on Intel TBB (Threading Building Blocks). On other platforms a less-performant fallback implementation based on the TinyThread library is used.

For additional documentation see the package website at: http://rcppcore.github.io/RcppParallel.

Author(s)

JJ Allaire, Romain Francois, Gregory Vandenbrouck, Marcus Geelnard, Intel Inc.

[Package RcppParallel version 4.4.1 Index]