RcppParallelFlags {RcppParallel}	R Documentation

Compilation flags for RcppParallel

Description

Output the compiler or linker flags required to build against RcppParallel.

Usage

CxxFlags()
LdFlags()
RcppParallelLibs()

Details

These functions are typically called from Makevars as follows:

PKG_LIBS += $(shell "${R_HOME}/bin/Rscript" -e "RcppParallel::LdFlags()")

Value

Returns NULL invisibly. The function is not called for it's return value rather for the side effect of outputting the flags.

[Package RcppParallel version 4.4.1 Index]